cli: extract perspective functions
