+++ cli/src/util.rs
@@ -76,7 +76,6 @@ pub async fn string_2_perspective_snapshot(
     cap_token: String,
     string: String,
 ) -> Result<Perspective> {
-    use crate::perspectives;
     use rand::distributions::Alphanumeric;
     use rand::{thread_rng, Rng};
 
@@ -85,9 +84,9 @@ pub async fn string_2_perspective_snapshot(
         .take(30)
         .map(char::from)
         .collect();
-    let temp_perspective = perspectives::run_add(cap_token.clone(), rand_name).await?;
+    let temp_perspective = ad4m_client::perspectives::run_add(cap_token.clone(), rand_name).await?;
     println!("Created temporary perspective: {}", temp_perspective);
-    perspectives::run_add_link(
+    ad4m_client::perspectives::run_add_link(
         cap_token.clone(),
         temp_perspective.clone(),
         "ad4m://self".to_string(),
@@ -97,10 +96,10 @@ pub async fn string_2_perspective_snapshot(
     .await?;
     println!("Added status link to temporary perspective");
 
-    let snapshot = perspectives::run_snapshot(cap_token.clone(), temp_perspective.clone()).await?;
+    let snapshot = ad4m_client::perspectives::run_snapshot(cap_token.clone(), temp_perspective.clone()).await?;
     println!("Created snapshot of temporary perspective");
 
-    perspectives::run_remove(cap_token.clone(), temp_perspective).await?;
+    ad4m_client::perspectives::run_remove(cap_token.clone(), temp_perspective).await?;
 
     Ok(snapshot)
 }
